Move mock_update_resource to mocks
